requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.3
annotated-types==0.7.0
audioread==3.0.1
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
easydict==1.13
efficientnet_pytorch==0.7.1
einops==0.8.0
eval_type_backport==0.2.2
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
grpcio==1.70.0
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
kaggle==1.7.4.2
kagglehub==0.3.10
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
munch==4.0.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
pretrainedmodels==0.7.4
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
segmentation_models_pytorch==0.4.0
SimpleITK==2.4.1
simsimd==6.2.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.9
tokenizers==0.15.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.13.1+cu113
tqdm==4.67.1
transformers==4.39.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
